+++ src/gen/fhsm_dispatch.c
@@ -546,6 +546,16 @@ __attribute__((weak)) fhsm_rv_t dispatch_rsa_oaep(
     return FHSM_RV_FUNCTION_FAILED;
 }
 
+__attribute__((weak)) fhsm_rv_t dispatch_rsa_pkcs(
+    unsigned long session, unsigned long key,
+    const void *params, size_t plen,
+    fhsm_slice_t in, uint8_t *out, size_t *outlen)
+{
+    (void)session; (void)key; (void)params; (void)plen;
+    (void)in; (void)out; (void)outlen;
+    return FHSM_RV_FUNCTION_FAILED;
+}
+
 __attribute__((weak)) fhsm_rv_t dispatch_rsa_pss(
     unsigned long session, unsigned long key,
     const void *params, size_t plen,
@@ -577,6 +587,16 @@ __attribute__((weak)) fhsm_rv_t dispatch_rsa_pss_sha384(
 }
 
 __attribute__((weak)) fhsm_rv_t dispatch_rsa_pss_sha512(
+    unsigned long session, unsigned long key,
+    const void *params, size_t plen,
+    fhsm_slice_t in, uint8_t *out, size_t *outlen)
+{
+    (void)session; (void)key; (void)params; (void)plen;
+    (void)in; (void)out; (void)outlen;
+    return FHSM_RV_FUNCTION_FAILED;
+}
+
+__attribute__((weak)) fhsm_rv_t dispatch_rsa_x509(
     unsigned long session, unsigned long key,
     const void *params, size_t plen,
     fhsm_slice_t in, uint8_t *out, size_t *outlen)